Refactored packages & updated build.xml test exclusion
