+++ .metadata/.plugins/org.eclipse.debug.core/.launches/Main.launch
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.jdt.launching.localJavaApplication">
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/RS Timer/src/com/minepop/talkar/timer/Main.java"/>
+<listEntry value="/RS Timer"/>
 </listAttribute>
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
-<listEntry value="1"/>
+<listEntry value="4"/>
 </listAttribute>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="com.minepop.talkar.timer.Main"/>
+<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="io.github.talkarcabbage.rstimer.Main"/>
 <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="RS Timer"/>
 </launchConfiguration>
