+++ build_info.properties
@@ -1,4 +1,4 @@
 #Sun, 05 Mar 2017 14:27:40 -0600
 build.major.number=0
-build.minor.number=14
-build.revision.number=22
+build.minor.number=15
+build.revision.number=0
