+++ src/io/github/talkarcabbage/rstimer/PeriodicTimer.java
@@ -1,4 +1,4 @@
-package com.minepop.talkar.timer;
+package io.github.talkarcabbage.rstimer;
 
 import java.util.HashMap;
 import java.util.Map;
